#6 Adding extracting service implementation
